split test_scraper_crops_sync_workflow.py into smaller files
